+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 **Release date: June 7th, 2018**
  - Added ability to limit to multiple species rather than just one.
  - Added small table so top 5 hits are easily navigable rather than showing the top hit only.
+ - Added expression data from the EMBL-EBI Expression Atlas.
  - Fixed multiple entries for PFAM/PROSITE causing issues with both the UI and links themselves.
  - Fixed linebreaks not being properly removed from old queries.
  - Added GO function and Interpro domains fields.